refactor(test): ワークアウトテストファイルを機能別に分割
- workouts.test.js を認証・バリデーション別にファイル分割
- workouts-auth.test.js: 認証関連テストケース
- workouts-validation.test.js: バリデーション関連テストケース
- テストの保守性と可読性を向上
